forgot password flow

+++ app/models/validation.rb
@@ -7,9 +7,7 @@ class Validation < ApplicationRecord
 
   def verify_existing_validation
     old = Validation.find_by(user_id: user_id)
-    unless old.nil?
-      old.destroy
-    end
+    old.destroy unless old.nil?
     self.expiration = DateTime.now + 1.hour
   end
 end
